resolve clang error and cspell
