+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_file_client.hpp
@@ -380,7 +380,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
         const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
-     * @brief NFS only. Creates a symoblic link to the file specified by path.
+     * @brief NFS only. Creates a symbolic link to the file specified by path.
      * @param linkText The absolution or relative path to the file to be linked to.
      * @param options Optional parameters to create this file's symbolic link.
      * @param context Context for cancelling long running operations.
